+++ src/index.ts
@@ -4,7 +4,16 @@ export { EventDispatcher, Event } from "./lib/utils/events";
 export { resetReflectionID } from "./lib/models/reflections/abstract";
 export { normalizePath } from "./lib/utils/fs";
 export * from "./lib/models";
-export { Converter, Context, type CommentParserConfig } from "./lib/converter";
+export {
+    Converter,
+    Context,
+    type CommentParserConfig,
+    type DeclarationReference,
+    type SymbolReference,
+    type ComponentPath,
+    type Meaning,
+    type MeaningKeyword,
+} from "./lib/converter";
 
 export {
     Renderer,