+++ src/lib/output/themes/default/DefaultThemeRenderContext.ts
@@ -74,8 +74,15 @@ export class DefaultThemeRenderContext {
         return md ? this.theme.markedPlugin.parseMarkdown(md) : "";
     };
 
-    attemptExternalResolution = (type: ReferenceType) => {
-        return this.theme.owner.attemptExternalResolution(type);
+    /**
+     * Using this method will repeat work already done, instead of calling it, use `type.externalUrl`.
+     * @deprecated
+     * Will be removed in 0.24.
+     */
+    attemptExternalResolution = (type: NeverIfInternal<ReferenceType>) => {
+        return this.theme.owner.attemptExternalResolution(
+            (type as ReferenceType).toDeclarationReference()
+        );
     };
 
     reflectionTemplate = bind(reflectionTemplate, this);